+++ lib/src/arg/use_key_name.dart
@@ -1,0 +1,16 @@
+import 'arg.dart';
+
+class UseKeyName extends Argument<bool> {
+  @override
+  String? get abbr => null;
+
+  @override
+  bool get defaultsTo => false;
+
+  @override
+  String get help =>
+      'Whether to use keyName style (packages/{package}/...) for asset paths';
+
+  @override
+  String get name => 'use-key-name';
+}